backend/pyproject.toml
[tool.poetry]
name = "heart-web"
version = "1.0.0"
description = "heart-web application"
authors = ["heart-web"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.2"
uvicorn = "^0.14.0"
stringcase = "^1.2.0"
SQLAlchemy = "^1.4.19"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
pytest = "^6.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
